requirements.txt
Flask
nltk
scikit-learn
python-docx
PyPDF2
gunicorn
